+++ cg/cli/set/family.py
@@ -1,9 +1,11 @@
 """Set case attributes in the status database"""
 import logging
+from typing import Optional, Tuple
 
 import click
-
 from cg.constants import CASE_ACTIONS, PRIORITY_OPTIONS, DataDelivery, Pipeline
+from cg.models.cg_config import CGConfig
+from cg.store import Store, models
 from cg.utils.click.EnumChoice import EnumChoice
 
 LOG = logging.getLogger(__name__)
@@ -29,20 +31,20 @@ LOG = logging.getLogger(__name__)
 @click.option("-g", "--panel", "panels", multiple=True, help="update gene panels")
 @click.option("-p", "--priority", type=click.Choice(PRIORITY_OPTIONS), help="update priority")
 @click.argument("family_id")
-@click.pass_context
+@click.pass_obj
 def family(
-    context: click.Context,
-    action: str,
-    data_analysis: Pipeline,
-    data_delivery: DataDelivery,
-    priority: str,
-    panels: [str],
+    context: CGConfig,
+    action: Optional[str],
+    data_analysis: Optional[Pipeline],
+    data_delivery: Optional[DataDelivery],
+    priority: Optional[str],
+    panels: Optional[Tuple[str]],
     family_id: str,
-    customer_id: str,
+    customer_id: Optional[str],
 ):
     """Update information about a case."""
-
-    case_obj = context.obj["status_db"].family(family_id)
+    status_db: Store = context.status_db
+    case_obj: models.Family = status_db.family(family_id)
     if case_obj is None:
         LOG.error("Can't find case %s,", family_id)
         raise click.Abort
@@ -53,7 +55,7 @@ def family(
         LOG.info("Update action: %s -> %s", case_obj.action or "NA", action)
         case_obj.action = action
     if customer_id:
-        customer_obj = context.obj["status_db"].customer(customer_id)
+        customer_obj: models.Customer = status_db.customer(customer_id)
         if customer_obj is None:
             LOG.error("Unknown customer: %s", customer_id)
             raise click.Abort
@@ -67,7 +69,7 @@ def family(
         case_obj.data_delivery = data_delivery
     if panels:
         for panel_id in panels:
-            panel_obj = context.obj["status_db"].panel(panel_id)
+            panel_obj: models.Panel = status_db.panel(panel_id)
             if panel_obj is None:
                 LOG.error(f"unknown gene panel: {panel_id}")
                 raise click.Abort
@@ -77,4 +79,4 @@ def family(
         LOG.info(f"update priority: {case_obj.priority_human} -> {priority}")
         case_obj.priority_human = priority
 
-    context.obj["status_db"].commit()
+    status_db.commit()